tests/examples/factory/requirements.txt
pathmind==0.3
ruamel.yaml>=0.17
pandas
xlrd
